+++ azure_devops_rust_api/examples/utils/mod.rs
@@ -1,7 +1,9 @@
-use azure_devops_rust_api::Credential;
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+use async_trait::async_trait;
+use azure_core::{Context, Policy, PolicyResult, Request};
+use azure_devops_rust_api::Credential;
 use azure_identity::{AutoRefreshingTokenCredential, DefaultAzureCredentialBuilder};
 use std::sync::Arc;
 
@@ -38,5 +40,30 @@ pub fn get_credential() -> Credential {
             Credential::from_pat(token)
         }
         _ => authenticate_with_default_credential(),
+    }
+}
+
+// Define a simple policy to always set the `accept` header to `application/zip`
+#[derive(Default, Debug, Clone, PartialEq, Eq)]
+pub struct AcceptZipPolicy {}
+
+impl AcceptZipPolicy {
+    #[allow(dead_code)]
+    pub fn new_policy() -> Arc<dyn Policy> {
+        Arc::new(AcceptZipPolicy {})
+    }
+}
+
+/// Always set the `accept` header to `application/zip`
+#[async_trait]
+impl azure_core::Policy for AcceptZipPolicy {
+    async fn send(
+        &self,
+        ctx: &Context,
+        request: &mut Request,
+        next: &[Arc<dyn Policy>],
+    ) -> PolicyResult {
+        request.insert_header("accept", "application/zip");
+        next[0].send(ctx, request, &next[1..]).await
     }
 }